feat(datum): constructor defaults
