+++ src/core/typesInternal.ts
@@ -1,4 +1,4 @@
-import { StoredVariant } from './types'
+import { DefaultsBase, InputBase, StoredVariant } from './types'
 import { z } from 'zod'
 
 export type SomeVariant = {
@@ -6,6 +6,9 @@ export type SomeVariant = {
   schema: z.SomeZodObject
   encode: never | SomeEncoderDefinition
   decode: never | SomeDecoderDefinition
+  _: {
+    defaultsProvider: null | SomeDefaultsProvider
+  }
 }
 
 export type SomeVariantConstructorInput = Record<string, unknown>
@@ -24,9 +27,15 @@ export type SomeDecoderDefinition = (
   extensions: { schema: SomeVariantSchema; name: string; [key: string]: unknown }
 ) => null | SomeVariantConstructorInput
 
-export type SomeVariantDefinition = Omit<StoredVariant, 'codec' | 'schema'> & {
+export type SomeDefaultsProvider<
+  PotentialInput extends InputBase = InputBase,
+  Defaults extends DefaultsBase = DefaultsBase
+> = (potentialInput: PotentialInput) => Defaults
+
+export type SomeVariantDefinition = Omit<StoredVariant, 'codec' | 'schema' | 'defaults'> & {
   codec?: SomeCodecDefinition
   schema: z.SomeZodObject
+  defaultsProvider: null | SomeDefaultsProvider
 }
 
 export type SomeADT = {
