+++ src/datum/builder.ts
@@ -1,14 +1,22 @@
 import { Initial } from './types'
 import { is } from '~/core/helpers'
 import { ExtensionsBase } from '~/core/types'
-import { SomeCodecDefinition, SomeSchema, SomeVariant, SomeVariantDefinition } from '~/core/typesInternal'
+import {
+  SomeCodecDefinition,
+  SomeDefaultsProvider,
+  SomeSchema,
+  SomeVariant,
+  SomeVariantDefinition,
+} from '~/core/typesInternal'
 import { z } from 'zod'
 
 export const datum = <Name extends string>(name: Name): Initial<Name> => {
+  const initialSchema = z.object({ _tag: z.literal(name) })
   const currentVariant: SomeVariantDefinition = {
     name,
-    schema: z.object({ _tag: z.literal(name) }),
+    schema: initialSchema,
     extensions: {},
+    defaultsProvider: null,
   }
 
   const builder = {
@@ -28,23 +36,34 @@ export const datum = <Name extends string>(name: Name): Initial<Name> => {
       currentVariant.codec = codecDef
       return builder
     },
+    defaults: (defaultsProvider: SomeDefaultsProvider) => {
+      if (currentVariant.schema === initialSchema) throw new Error(`No schema defined.`)
+      if (currentVariant.defaultsProvider) throw new Error(`Defaults already defined.`)
+      currentVariant.defaultsProvider = defaultsProvider
+      return builder
+    },
     done: () => {
       const symbol = Symbol(currentVariant.name)
       const controller = {
         ...currentVariant,
+        _: {
+          defaultsProvider: currentVariant.defaultsProvider,
+        },
         create: (input?: object) => ({
           _tag: currentVariant.name,
           _: {
             symbol,
           },
-          ...input,
+          // TODO pass through zod validation
+          ...applyDefaults(input ?? {}, currentVariant.defaultsProvider?.(input ?? {}) ?? {}),
         }),
+        // TODO move into _
         symbol,
         //eslint-disable-next-line
         is$: (value: unknown) => is(value, symbol),
         is: (value: unknown) => is(value, symbol),
         decode: (value: string) => {
-          if (!currentVariant.codec) throw new Error(`Codec not implemented.`)
+          if (!currentVariant.codec) throw new Error(`Codec not defined.`)
           const data = currentVariant.codec.decode(value, {
             ...currentVariant.extensions,
             schema: currentVariant.schema,
@@ -68,6 +87,16 @@ export const datum = <Name extends string>(name: Name): Initial<Name> => {
       }
       return controller
     },
+  }
+
+  const applyDefaults = (input: object, defaults: object) => {
+    const input_ = { ...input }
+    for (const entry of Object.entries(defaults)) {
+      // @ts-expect-error dynammic
+      // eslint-disable-next-line
+      input_[entry[0]] = input_[entry[0]] === undefined ? entry[1] : input_[entry[0]]
+    }
+    return input_
   }
 
   // TODO
