+++ src/datum/types.ts
@@ -2,14 +2,14 @@
  * This module is concerned with the static types for the API of building up an ADT.
  */
 
-import { z } from '../lib/z'
 import { Datum } from './controller'
-import { CodecDefiniton, ExtensionsBase, NameBase, SchemaBase, StoredVariant } from '~/core/types'
+import { CodecDefiniton, ExtensionsBase, SchemaBase, StoredVariant } from '~/core/types'
+import { SomeDefaultsProvider } from '~/core/typesInternal'
 
 /**
  * The initial API for building an ADT.
  */
-export type Initial<Name extends string> = PostName<CreateStoredVariant<Name>>
+export type Initial<Name extends string> = PostName<StoredVariant.Create<Name>>
 
 /**
  * The builder API when it is in a state where a variant is required.
@@ -33,16 +33,22 @@ export interface PostName<V extends StoredVariant> extends Done<V> {
  */
 export interface PostSchema<V extends StoredVariant> extends Done<V> {
   codec(definition: CodecDefiniton<V>): PostCodec<StoredVariant.AddCodec<V>>
-  /**
-   * Extend the ADT with new properties.
-   * TODO
-   */
+  // prettier-ignore
+  defaults<Defaults extends Partial<StoredVariant.GetType<V>>>(defaults: SomeDefaultsProvider<Partial<StoredVariant.GetType<V>>, Defaults>): PostDefaults<StoredVariant.AddDefaults<V, Defaults>>
   // prettier-ignore
   extend<Extensions extends ExtensionsBase>(extensions: Extensions): PostExtend<StoredVariant.AddExtensions<Extensions, V>>
 }
 
 export interface PostExtend<V extends StoredVariant> extends Done<V> {
+  // prettier-ignore
+  defaults<Defaults extends Partial<StoredVariant.GetType<V>>>(defaults: SomeDefaultsProvider<Partial<StoredVariant.GetType<V>>, Defaults>): PostDefaults<StoredVariant.AddDefaults<V, Defaults>>
   codec(definition: CodecDefiniton<V>): PostCodec<StoredVariant.AddCodec<V>>
+}
+
+export interface PostDefaults<V extends StoredVariant> extends Done<V> {
+  codec(definition: CodecDefiniton<V>): PostCodec<StoredVariant.AddCodec<V>>
+  // prettier-ignore
+  extend<Extensions extends ExtensionsBase>(extensions: Extensions): PostExtend<StoredVariant.AddExtensions<Extensions, V>>
 }
 
 export type PostCodec<V extends StoredVariant> = Done<V>
@@ -65,15 +71,6 @@ export interface Done<V extends StoredVariant> {
 }
 
 // Helpers
-
-export type CreateStoredVariant<Name extends NameBase> = {
-  name: Name
-  schema: { _tag: z.ZodLiteral<Name> }
-  codec: false
-  // TODO
-  // eslint-disable-next-line
-  extensions: {}
-}
 
 export type StoredDatum<Name extends string = string> = {
   name: Name
